Fixed betty error

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,16 +11,17 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	int j;
+	unsigned int j;
 
 	if (array == NULL)
 		return (-1);
 
-	for (j = 0; j < (int) size; j++)
+	for (j = 0; j < size; j++)
 	{
 		printf("Value checked array[%d] = [%d]\n", j, array[i]);
 		if (array[j] == value)
 			return (j);
 	}
+
 	return (-1);
 }
